backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0
joblib==1.4.2
scikit-learn==1.5.1
xgboost==2.1.1
pandas==2.2.2
numpy==1.26.4
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0